+++ src/app/projects/page.tsx
@@ -1,9 +1,13 @@
-import ProjectsImpl from "@/components/org/jallen/implementation/ProjectsImpl";
+import { getApplications } from '../data/api/application';
+import ProjectsImpl from '@/components/org/jallen/implementation/ProjectsImpl';
+import { Application, ProcessWork } from '../data/common/types';
+import { getProcessWorks } from '../data/api/work';
 
-
-const Projects = () => {
+const Projects = async () => {
+  const applications: Application[] = await getApplications();
+  const works: ProcessWork[] = await getProcessWorks();
   return (
-    <ProjectsImpl />
+    <ProjectsImpl applications={applications} works={works} />
   )
 }