+++ src/app/resume/page.tsx
@@ -1,8 +1,17 @@
-import ResumeImpl from "@/components/org/jallen/implementation/ResumeImpl";
+import { About, Experience, Information, Skills } from '../data/common/types';
+import { getAbouts } from '../data/api/about';
+import { getExperiences } from '../data/api/experience';
+import { getSkillsWithIcons } from '../data/api/skills';
+import ResumeImpl from '@/components/org/jallen/implementation/ResumeImpl';
+import { getInformations } from '../data/api/information';
 
-const Resume = () => {
+const Resume = async () => {
+  const abouts: About[] = await getAbouts();
+  const experiences: Experience[] = await getExperiences();
+  const skills: Skills[] = await getSkillsWithIcons();
+  const informations: Information[] = await getInformations();
   return (
-    <ResumeImpl />
+    <ResumeImpl abouts={abouts} experiences={experiences} skills={skills} informations={informations} />
   )
 }