+++ src/components/org/jallen/implementation/ResumeImpl.tsx
@@ -19,63 +19,8 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { motion } from 'framer-motion';
-
-const general = {};
-
-// details: [
-//   {
-//     fieldName: 'Name',
-//     fieldValue: 'James Tadique'
-//   },
-//   {
-//     fieldName: 'Phone',
-//     fieldValue: '0000'
-//   },
-//   {
-//     fieldName: 'Experience',
-//     fieldValue: '11 Years'
-//   },
-//   {
-//     fieldName: 'Email',
-//     fieldValue: 'name@yahoo.com'
-//   },
-//   {
-//     fieldName: 'Email II',
-//     fieldValue: 'name@gmail.com'
-//   },
-//   {
-//     fieldName: 'Nationality',
-//     fieldValue: 'Filipino'
-//   }
-// ]
-
-
-const experience = {
-  title: 'Experience',
-  description: 'Description',
-  details: [
-    {
-      company: 'Company',
-      position: 'Position',
-      duration: '1'
-    },
-    {
-      company: 'Company',
-      position: 'Position',
-      duration: '2'
-    },
-    {
-      company: 'Company',
-      position: 'Position',
-      duration: '3'
-    },
-    {
-      company: 'Company',
-      position: 'Position',
-      duration: '4'
-    }
-  ]
-};
+import { About, Experience, Heading, Information, Skills } from '@/app/data/common/types';
+import { getProfileHeading, getExperienceHeading, getSkillsHeading } from '@/app/data/api/common';
 
 const skills = {
   title: 'Skills',
@@ -115,12 +60,15 @@ const skills = {
     },
     {
       icon: <SiNextdotjs />,
-      level: 'Next.js'
+      name: 'Next.js'
     }
   ]
 };
 
-const ResumeImpl = () => {
+const ResumeImpl = ({ abouts, experiences, skills, informations } : { abouts: About[], experiences: Experience[], skills: Skills[], informations: Information[] }) => {
+  const profileHeading: Heading = getProfileHeading();
+  const experienceHeading: Heading = getExperienceHeading();
+  const skillsHeading: Heading = getSkillsHeading();
   return (
     <motion.div
       initial={{ opacity: 0 }}
@@ -135,53 +83,73 @@ const ResumeImpl = () => {
       className='min-h-[80vh] flex items-center justify-center py-12 xl:py-0 gap-6'
     >
       <div className='container mx-auto'>
-        <Tabs defaultValue='experience' className='flex flex-col xl:flex-row gap-[60px]'>
+        <Tabs defaultValue='profile' className='flex flex-col xl:flex-row gap-[60px]'>
           <TabsList className='flex flex-row w-full max-w-[380px] mx-auto xl:flex-col xl:mx-0 bg-transparent'>
-            <TabsTrigger value='profile'>Profile</TabsTrigger>
-            <TabsTrigger value='experience'>Experience</TabsTrigger>
-            <TabsTrigger value='skills'>Skills</TabsTrigger>
+            <TabsTrigger value={profileHeading.value}>{profileHeading.title}</TabsTrigger>
+            <TabsTrigger value={experienceHeading.value}>{experienceHeading.title}</TabsTrigger>
+            <TabsTrigger value={skillsHeading.value}>{skillsHeading.title}</TabsTrigger>
+
+            <div className='hidden text-white m-6 ps-[20px] xl:block'>
+              {
+                informations.map((item, index) => {
+                  return (
+                    <div 
+                      key={index}
+                      className='flex flex-row w-full justify-between'
+                    >
+                      <div className='text-accent'>{item.name}:</div>
+                      <div className='text-white/60 font-ruka'>{item.value}</div>
+                    </div>
+                  )
+                })
+              }
+            </div>
           </TabsList>
 
           <div className='min-h-[70vh] w-full'>
-            <TabsContent value='profile' className='w-full'>
+            <TabsContent value={profileHeading.value} className='w-full'>
               <div className='flex flex-col gap-[30px]'>
                 <div className='flex flex-col gap-[10px] text-center xl:text-left'>
-                  {/* <h3 className='text-4xl font-bold'>{profile.title}</h3> */}
-                  {/* <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{profile.description}</p> */}
+                  <h3 className='text-4xl font-bold'>{profileHeading.title}</h3>
+                  <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{profileHeading.description}</p>
                 </div>
 
                 <ScrollArea className='h-[400px]'>
-                  <div className='flex flex-col gap-4'>
-                    <p>I am a full-stack developer with 11 years of total experience in different technical stacks. A person that finds enjoyment in learning about development such as different languages or technology, open to new challenges in coding and can be a team player. I thrive in research and development while delivering high quality systems.</p>
-
-                    <p>I am skilled at different libraries and frameworks such as JavaScript, TypeScript, React, Next.js and Angular using Node.js for server environment for front-end development. For back-end development, I am adept at using Java with knowledge of Spring framework, Spring boot, OAuth, Maven with great understanding of web services, microservices and restful API.</p>
-
-                    <p>I have good experience in using database management system such as MySQL, PostgreSQL, NoSQL, SQLite and Oracle Database. I am also proficient in using version control system such as Git with added fluency in deployment using Jenkins, Railway and Vercel.</p>
-
-                    <p>Additionally, I have knowledge with Python in small scale projects. I am also adept at creating test scripts using Playwright and Puppeteer to capture regression bugs that will occasionally happen during development cycle.</p>
+                  <div className='flex flex-col gap-2 bg-gray-700 rounded-xl'>
+                    {
+                      abouts.map((item: About, index: number) => {
+                        return (
+                          <p className='p-6' key={index}>{item.description}</p>
+                        )
+                      })
+                    }
                   </div>
                 </ScrollArea>
               </div>
             </TabsContent>
 
-            <TabsContent value='experience' className='w-full'>
+            <TabsContent value={experienceHeading.value} className='w-full'>
               <div className='flex flex-col gap-[30px] text-center xl:text-left'>
-                <h3 className='text-4xl font-bold'>{experience.title}</h3>
-                <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{experience.description}</p>
+                <div className='flex flex-col gap-[10px] text-center xl:text-left'>
+                  <h3 className='text-4xl font-bold'>{experienceHeading.title}</h3>
+                  <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{experienceHeading.description}</p>
+                </div>
                 <ScrollArea className='h-[400px]'>
-                  <ul className='grid grid-cols-1 lg:grid-cols-2 gap-[30px]'>
+                  <ul className='flex flex-col gap-4'>
                     {
-                      experience.details.map((item, index) => {
+                      experiences.map((item: Experience, index: number) => {
                         return (
                           <li 
                             key={index}
-                            className='bg-gray-700 h-[184px] py-6 px-10 rounded-xl flex flex-col justify-center items-center lg:items-start gap-1'
-                          >
-                            <span className='text-accent'>{item.duration}</span>
-                            <h3 className='text-xl max-w-[260px] min-h-[60px] text-center lg:text-left'>{item.position}</h3>
-                            <div>
-                              <span></span>
-                              <p>{item.company}</p>
+                            className='bg-gray-700 h-auto py-6 px-10 rounded-xl flex flex-col justify-center items-center lg:items-start gap-1'
+                          > 
+                            <h3 className='text-xl font-semibold max-w-[400px] min-h-[20px] text-center lg:text-left'>{item.company}</h3>
+                            <h4>{item.position}</h4>
+                            <span className='text-accent text-sm'>{item.duration}</span>
+                            <span className='text-white/60 text-sm'>{item.total}</span>
+                            <span className='text-white/60 text-sm'>{item.location}</span>
+                            <div className='mt-2'>
+                              <p>{item.description}</p>
                             </div>
                           </li>
                         )
@@ -192,33 +160,37 @@ const ResumeImpl = () => {
               </div>
             </TabsContent>
 
-            <TabsContent value='skills' className='w-full h-full'>
+            <TabsContent value={skillsHeading.value} className='w-full h-full'>
               <div className='flex flex-col gap-[30px]'>
-                <div className='flex flex-col gap-[30px] text-center xl:text-left'>
-                  <h3 className='text-4xl font-bold'>{skills.title}</h3>
-                  <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{skills.description}</p>
+                <div className='flex flex-col gap-[10px] text-center xl:text-left'>
+                  <h3 className='text-4xl font-bold'>{skillsHeading.title}</h3>
+                  <p className='max-w-[600px] text-white/60 mx-auto xl:mx-0'>{skillsHeading.description}</p>
                 </div>
-
-                <ul className='grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 gap-4 xl:gap-[30px]'>
-                  {
-                    skills.details.map((item, index) => {
-                      return (
-                        <li key={index}>
-                          <TooltipProvider delayDuration={100}>
-                            <Tooltip>
-                              <TooltipTrigger className='w-full h-[150px] bg-gray-700 rounded-xl flex justify-center items-center group'>
-                                <div className='text-6xl group-hover:text-accent transition-all duration:300'>{item.icon}</div>
-                              </TooltipTrigger>
-                              <TooltipContent>
-                                <p>{item.name}</p>
-                              </TooltipContent>
-                            </Tooltip>
-                          </TooltipProvider>
-                        </li>
-                      )
-                    })
-                  }
-                </ul>
+                <ScrollArea className='h-[400px]'>
+                  <ul className='grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 gap-4 xl:gap-[30px]'>
+                    {
+                      skills.map((item: Skills, index: number) => {
+                        if (item.icon !== undefined) {
+                          return (
+                            <li key={index}>
+                              <TooltipProvider delayDuration={100}>
+                                <Tooltip>
+                                  <TooltipTrigger className='w-full h-[150px] bg-gray-700 rounded-xl flex justify-center items-center group'>
+                                    <div className='text-6xl group-hover:text-accent transition-all duration:300'>{item.icon}</div>
+                                  </TooltipTrigger>
+                                  <TooltipContent>
+                                    <p>{item.name}</p>
+                                  </TooltipContent>
+                                </Tooltip>
+                              </TooltipProvider>
+                            </li>
+                          )
+                        }
+                      })
+                    }
+                  </ul>
+                </ScrollArea>
+                
               </div>
             </TabsContent>
           </div>