+++ src/components/org/jallen/socials/Socials.tsx
@@ -1,16 +1,6 @@
+import { getSocials } from '@/app/data/api/common';
+import { Social } from '@/app/data/common/types';
 import Link from 'next/link';
-import { FaGithub, FaLinkedin } from 'react-icons/fa';
-
-const socials = [
-  {
-    icon: <FaGithub />,
-    path: 'https://github.com/tjamesallen15',
-  },
-  {
-    icon: <FaLinkedin />,
-    path: 'https://www.linkedin.com/in/tjamesallen15/',
-  },
-];
 
 type SocialsProps = {
   containerStyles: string,
@@ -18,13 +8,14 @@ type SocialsProps = {
 };
 
 const Socials: React.FC<SocialsProps> = ({ containerStyles, iconStyles }) => {
+  const socialList: Social[] = getSocials();
   return (
     <div className={containerStyles}>
       {
-        socials.map((item, index) => {
+        socialList.map((item, index) => {
           return (
             <Link
-              href={item.path}
+              href={item.href}
               key={index}
             >
               <div className={iconStyles}>