+++ CHANGELOG.md
@@ -1,29 +1,43 @@
-# What's New with Jallen-Portfolio
+# What's New with Jallen-Next
 
 All notable changes to this project will be documented in this file.
 
 # 1.1.0
 
+## 1.1.0 Features
+
+`[About]` Added fetch calls for about data.
+`[Application]` Added fetch calls for application data.
+`[Common]` Added fetch calls for common data.
+`[Experience]` Added fetch calls for experience data.
+`[Information]` Added fetch calls for information data.
+`[Skills]` Added fetch calls for skills data.
+`[Work]` Added fetch calls for work data.
+
 ## 1.1.0 Fixes
 
-`[Header]` Fixed design and styles.
 `[Footer]` Fixed design and styles.
+`[Header]` Fixed design and styles.
+`[Index]` Added data fetch for skills badges.
 `[MobileNavigation]` Fixed design and styles.
 `[Navigation]` Fixed design and styles.
 `[Profile]` Fixed design and styles.
-`[Socials]` Fixed design and styles.
 `[Projects]` Fixed design and styles.
+`[Projects]` Added data fetch calls.
 `[Resume]` Fixed design and styles.
+`[Resume]` Added data fetch calls.
+`[Socials]` Fixed design and styles.
+`[Types]` Fixed type file structure.
 
 # 1.0.0
 
 ## 1.0.0 Features
 
-`[Header]` Added header component.
 `[Footer]` Added footer component.
+`[Header]` Added header component.
 `[MobileNavigation]` Added mobile navigation component.
 `[Navigation]` Added navigation component.
 `[Profile]` Added profile component.
-`[Socials]` Added socials component.
 `[Projects]` Added projects component.
 `[Resume]` Added resume component.
+`[Socials]` Added socials component.